Agregado sweetAlert y reactSnippers

+++ src/components/CartContainer/CartContainer.jsx
@@ -1,4 +1,5 @@
 import React, { useContext } from 'react'
+import Swal from 'sweetalert2'
 import { CartContext } from '../../context/CartContext'
 import Cart from './Cart'
 
@@ -16,7 +17,24 @@ const CartContainer = () => {
   }
 
   const handleDeleteProduct = (id) => {
-    return deleteProduct(id)
+    Swal.fire({
+      title: 'Estas seguro que deseas eliminar el producto del carrito?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Eliminado',
+          'Se eliminó correctamente el producto del carrito',
+        )
+        return deleteProduct(id)
+      }
+    })
+   
   }
 
   const totalAmount = getCartTotalAmount();
@@ -24,7 +42,23 @@ const CartContainer = () => {
   const totalQuantity = getCartTotalQuantity();
 
   const handleCleanCart = () => {
-    return cleanCart()
+    Swal.fire({
+      title: 'Estas seguro que deseas limpiar el carrito?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Eliminado',
+          'Se limpió correctamente todo el carrito',
+        )
+        return cleanCart()
+      }
+    })
   }
